requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.7.4
python-dotenv==1.0.1
openai==1.35.0
httpx==0.27.0